style: readability improvements on mobile
Previously, I was blanket "max-width: 60%" to center the posts. It is
cleaner to use the @media selector to do this conditionally based on
screen size, as we actually don't want the width constraint on mobile.

Also, use "vmin" font-sizes: it uses the "virtual window" to compute the
font size. It is a 3 multiplier by the min(height,width) of the window.
Cool, huh?