style: found solution to mobile text size
I finally found the solution to the proper mobile text zooming! The
solution was to use the "meta viewport" tag, which seems to detect the
small device and scale up the font. After scouring the CSS docs, the
solution was an HTML tag all along... go figure!

More on this here:

    https://developer.mozilla.org/en-US/docs/Mozilla/Mobile/Viewport_meta_tag