thoughts: Add post on Domain Names
This was super cathartic to write. I've wasted a lot of time this week
pouring over domain registrars and learning about the whole system and
why it is:

    1) (Almost) Impossible to get a domain you want
    2) (Almost always) Impossible to get a domain you want cheap

Seriously. 'Premium Fees' make me sick -- and looking at all the
old-timers on the internet, just makes me wonder what they paid for that
slick .org.

Sigh. Feels good to shame RealNames a bit. I mean that was the _worst_.
Want to rent the domain for your website / email? We'll give you the
latter for 3x the price :(

roar