make: Update install recipe for OpenBSD host
I now host on OpenBSD, so the destination is in a different place.

It's been a long road to OpenBSD, full of hardware problems (hot
laptops, X hanging on my amdgpu, etc), but I really admire the
security-first aspect of the OS, along with the simplicity of the base
install.

I'm a huge believer in simple software being better for everyone
(maintainers, end users, developers, sysadmins...), and httpd+relayd
is a nice stack that I'm excited to host on.

Maybe I'll add a note about it on the "thoughts" page :)