templatize: Split thoughts into posts, templatize everything
This was supposed to be two commits, one for the build system and one
for splitting up thoughts.html ... but it appears something went wrong
and nvi aborted without committing :( and now the effort is too high to
undo it.

So without further ado, I present: massive refactor commit! This adds a
brand new templating script (if you can even call it that...) tm.pl. I
seriously considered hugo and other solutions, but the temptation to
write my own was just too strong. I don't need a feature heavy library
and I don't really want that dependency. 35 lines of Perl is light
enough for now, and it's a great learning experience. Also, tm.pl is
just too clever to pass up :)

So on top of now needing to build the static content, this commit
refactored thoughts into a series of posts, each with their own
head/tail template.

This feels like the right scalability solution, in the sense that
updating the style of many posts will only involve updating the
templates.

Likewise, having one long "thoughts" page is both unlinkable (and thus
unshareable) and just not practical.

So thoughts blooms into a maybe-one-day blog.

Cheers!