style: Replace /home/akarle with real header
This commit replaces the temporary /home/akarle header with a more
appealing/clean nav bar that has my name in it.

That way, you know who you're hearing from, and there isn't this obscure
nerdy reference at the top of the site.

Used media queries to stack the flex box when it gets too small. Looks
nice on both mobile and desktop!