thoughts: Add Black Lives Matter post
I don't want this to come off as about me -- it isn't. I thought of
writing about my experience going to a protest, but it felt too much
like I was trying to prove to be a "good guy".

Nevertheless, I think it's important to show support. If one day someone
does find this blog, I want to be on the right side of history.