content: Migrate to Markdown posts on top of templating script
Previously, I had whipped up a small templating script, tm.pl, which
would look for <!-- {% include=foo %} --> comments and expand them. This
allowed me to make edits to the general style of the site without
changing all the content.

However, I begrudgingly accepted that I needed a faster way to write
posts. Hand writing HTML is quite a pain... so I gave up some of my
stylistic freedoms that come with writing HTML and added the original
Markdown.pl script to the build pipeline (p5-Text-Markdown package on
OpenBSD).

Now, the markdown is translated to HTML before the tm.pl step, giving me
the best of both worlds. The only minor sacrifice was the "on hover"
underlines in the Thoughts page. Maybe I'll add them back, but it
doesn't look awful as is.