thoughts: Add post on migrating to self-hosting
I've been wanting to write this for a long time. For one, I wanted to
set the record straight that no, I was _NOT_ using Fastmail's "files"
GUI anymore (I mean... it worked, but it was painful and not
particularly technically impressive!)

And it felt like a necessary followup to the ramblings on domains. Yes,
I settled on alexkarle.com, and yes I've happily moved on to bigger and
better things (like self-hosted git repos! how awesome is that!)

I'd love to dive deeper into the topic, but I wanted to write at least
something so that my future self (AFAIK, the only reader) can look back
and say "wow! I remember those days"

Cheers