style: Move from monospace-everywhere to monospace-headers
While I originally liked the old-timey feel of being monospace
everywhere, I now subscribe to the ideology that it is harder to read
prose in monospace (this became most noticeable with my recent post that
was about 3x longer than any prior post...).

I like the monospace vibe in the headers and nav (for spacing reasons),
but now that I'm rendering / componsing in markdown, I needed a way to
signify the <code> blocks from the regular text.

Changing from monospace -> sans-serif makes it both easier to read and
easier to distinguish the <code> from the <p>.