thoughts: Add post on blogging into the void
As a non-technical post, there's not much backstory to put
in the commit message... the post I was referring to in the
beginning is a post on ed(1) and my discovery/learning/using
it over vacation. I do still plan on posting that one day...
just one more proof read, I promise! :P

This felt cathartic to write. Despite the late-hour, I'm gonna
push it. I fear if I wait until morning to re-read it, I won't
publish it at all.