nav: Update Git link git.alexkarle.com -> alexkarle.com/git/
Piloting this change. One of the cool advantages of using stagit(1) is
that the files truly are static HTML and so they can be put in a
subdirectory of the regular webserver's static content.

This just makes the transition clear that you aren't changing servers,
you're still talking to alexkarle.com, etc. I think that's a nicer flow.

That doesn't mean I'm doing away with git.alexkarle.com though! I think
it's important to use the git. subdomain in clone URLs so that I can
freely migrate the git hosting *withought* migrating the rest of the
server.