www: Goodbye markdown, long live mdoc
Man (pun intended), I had a blast putting this together. I got the
idea the other day while playing around with a toy static site, and
it just clicked--what if I used mdoc as my markup language and mandoc
to generate the HTML for a static site?

This was obviously not the intended use case, but I think it gives off
a nerdy vibe in exactly the right way :)

Later versions of mandoc allow .Xr to link against a fallback if the
reference doesn't exist locally, which is perfect for using .Xr as both
my internal site links (between pages) and as a cheap reference to
OpenBSD man pages (which I plan to do more often as I write more
technical content).

The beauty of this setup is that it can now be built entirely from
base OpenBSD (relying only on make(1) and mandoc(1)).

Time for the final test -- the girlfriend approval!