www: Change generated .7 files to .html (instead of .7.html)
I'm not 100% sure why I settled on .7.html. I think it made sense at the
time because it was the fastest way to adapt build.sh (prior to writing
the Makefile recipe), and it was the way man.openbsd.org did it.

But in reality, I don't plan on using any extension other than .7 (and
if I did the chance of a name conflict is low). I think dropping the .7
from the extension just makes the URL much cleaner looking!