cleanup: De-markdown-ify the README
As the final nail in the markdown coffin, I've moved README.md to plain
README, no markdown.

It's not that I have a personal vendetta against markdown (although...
if you've ever tried to render nested code blocks under nested lists...
hoooeee!). It's just that I get a small joy out of knowing that this
file won't play nicely with Windows ;)