make: Add atom feed to the build process!
This was a really fun sunday-morning activity. I learned a bit about the
Atom syndication format, and had some fun scripting up a simple generator
to take blog.7 and output a single atom.xml feed with the full content of
all articles in it!

I recently started using RSS/Atom, and I think it's a great alternative to
link aggregation sites (and social media)--like most things with this site,
I try to model the web I would love to see, so I had to put my money where
my mouth was and add atom to alexkarle.com :)