gopher: Add small mdoc -> gopher generation script
In my time off this week, I took a deep dive on gopher, and it's a
really cool niche blogging space! I intend to host all my posts there,
which is pretty easy because mdoc(7) translates nicely to ascii with
mandoc(1).

This patch adds a script to generate a *very* primitive gopherhole for
use with gophernicus. The key elements are:

1. Generates a TXT file for each blog post / page
2. Moves the blogs to phlog/
3. Sets the mtime of the blog posts so that the default DIR view shows
   them

I hope to expand on this in future commits by creating a gophermap for
the blog(7) page with links to the content (rather than just using the
DIR view). But for now... we're live!