general: License code and content appropriately
I've been meaning to do this for a long time, but kept waffling over the
choice of license, as well as how to display it. I wanted something that
would be easily found (i.e. on each page), but also that didn't detract
from the man page aesthetic ;)

I settled on a small footer just stating the copyright and linking to a
license file. A quick survey of my girlfriend showed that simply putting
"Content CC BY 4.0, Code MIT" at the bottom didn't make sense, and I
didn't want to wrap to multiple lines on mobile... so here we have it.

Plus, if I ever get around to creating an SSH-kiosk, it'll be good to
have it as a standalone page!