kiosk: Add MVP ssh kiosk program to browse blog via ssh+mandoc
I'm really excited (if a bit nervous about the security implications)
about this kiosk program. I've always thought it'd be cool to browse my
site via mandoc(1), so I wrote a simple shell-like program called
'kiosk' that lists the man pages in the site and asks which you'd like
to read (with some other options to quit, list, or see help).

This isn't 100% ready to go live (as a public SSH option), but it's at a
state where it's worth checking in!