jam-tuesday: Create HTML rollup of stats ("Archive")
This is really just a pretty HTML file that replaces the httpd
auto-index of the jam-tuesday directory.

Some cleanup may be in need later (piping the jam-stats.sh output to sed
is a bit ugly... likely better to teach it to emit HTML).

In general, this whole "jam-tuesday" project has had me questioning the
value of mdoc(7) for non blog-post things... I want to ensure the site
is accessible via WWW (with/without JS), Gopher (plaintext version
needed) and soon Gemini.. so it's unclear if I'd be better off moving to
a more plaintext friendly syntax (like back to markdown :/ ).