kiosk: Replace mandoc(1) usage with less(1) on prebuilt files
Here we go again with runtime optimization at the cost of build time!
This patch removes the usage of `mandoc -l` on the contents of the
site and replaces it with straight less(1).

This is done by generating the kiosk-facing content at build time.

The *real* win here is that it allows us to add unveil(2) calls to
the kiosk, hiding all but the exported files from the field of view
of the process!

I was _tempted_ to replace less(1) with just straight printing the
file line by line to stdout... but I thought that was a bit drastic!

It would give us a super tight pledge though... hmm