kiosk: Demo super secure pledge (stdio + rpath only)
This is mostly a "spike" for fun, but it's worth keeping in the git
history if I want to go back to it.

The idea here is that kiosk will be a login shell open to anonymous,
untrusted, users. As such, I want to be SUPER sure I don't give anyone
shell access to my machine. In this iteration of kiosk, it restricts its
view to just the kiosk/ directory (and /dev/tty for prompting), and then
immediately drops privelages to just standard io and read-only access to
those limited paths. This means that even if there were a bad use of
memory or something like that, the attacker can never spawn a new
process, talk over the internet, or do anything else nefarious to hurt
the system.

The cost? Well, not having the pager ability is kind-of a pain,
especially for longer blog posts... As such, I'll likely just revert
this patch in a commit or two (maybe I'll try writing my own pager.
hmm).