jam-tuesday: Add a list of all songs to the archive
I was looking for this information recently when jamming with a new
friend and trying to find a common intersection of music. It woulda
been really helpful to show all songs, not just the top hits (since
half of our artists are really obscure :shrug:), so this patch adds
just that!

Also fix a couple mistyped songs/artists.