blog: Add new entry on creative limitation in coding
Honestly, I'm putting this one out there because its been a while
since I've written, and I'd like to write at a more frequent
cadence. It seemed fairly non-polarizing (see fear-of-judgement in
on-writing(7)), and it was kinda a "ah-ha" moment when it clicked
that this was one of my favorite ways to learn tech!