text: Goodbye kiosk and gengopher, hello text.alexkarle.com
I'll never feel fully comfortable rolling out kiosk(1) via SSH,
and gengopher was always a hack (with the setting of mtimes and
all that). Yet, I've always wanted my site to be available not
just as HTML but in a more man(1)-y format.

After reading text.causal.agency, I knew that a ascii dump was
the way to go. It took me quite a while to get it all working in
a gmake and OpenBSD-make way, but I'm really happy with the
results!