mdoc: Convert phony Xr's from Lk to Xr via postprocessing
Ever since starting to dump the raw ascii of the manpages for
text.alexkarle.com, I've been aware of the ugliness coming from
out-of-base cross references (Xrs). For instance, I want to link
vim.org in my uses(7) page, so I previously had a .Lk element
to enable the inline link in the HTML:

    .Lk https://www.vim.org vim(1)

This "phony Xr" looks great in HTML but is super ugly in the ascii:

    vim(1): https://www.vim.org

I hypothesize that anyone actually reading the ascii probably knows
what Vim is and would much rather just see "vim(1)" -- there's no
hyperlinking on text.alexkarle.com anyways, so why bother posting
the URL? It just hurts the flow of the paragraph.

To solve this "links in HTML but just plain Xr in text", I added
a new postprocessing step to genpost.sh that uses sed(1) to change
any of the .Xr links that are invalid urls (not in base) to a known
good link via the LINKS file. I'm pretty pleased with how the sed(1)
command is genated using awk(1) and eval'd on the fly :)

The ascii dump doesn't care what's in base and what's not, so
no need for change there!