text-only: Add new blog post announcing text.alexkarle.com
This was a fun blog post to write. It started off hyper technical,
discussing the way I was using timestamp files to circumnavigate the
no-subdirectory limitation of the inference rules... and of course while
writing it I realized I should just replace it with an install target.

So I had to scrap half of what was written and rewrite it. And in doing
so I realized the technical bits weren't nearly as interesting as why I
would ever bother doing this in the first place, so this post became
much more WHY and less HOW.