make: Replace timestamp solution with install recipe
Previous to this commit, I was using timestamp files to represent the
build of each *.txt/*.html combo. This is in itself OK, but the other
weird thing going on was that `make build` not only built the files but
was trying to place the files in the perfect place. This led to a lot of
small recipes that amounted to "take this file and copy it into text/".

I realized that the much cleaner way to do placement was to add another
"install" recipe. Rather than care about the layout of the build, we can
just use `install` to place everything in the DESTDIR in the right spot.

With this change, there's no longer the need for individual "copy this
file to this dir" recipes (that's all in install) and we can split the
one "timestamp covers both targets" to two inference rules--one for
*.html and one for *.txt. This is much cleaner (the reason we couldn't
do it before was that we wanted to build into a subdirectory--we now
don't care since we install into a subdirectory).

It only took me 1hr writing a blog post about _how_ I did it to realize
I should do it differently... but better late than never!