blog: Add new entry on `make obj`; what, how, why
This is kind of an immediate followup to text-only(7),
which largely skimmed over the HOW and focused more on
just announcing text.alexkarle.com.

I'm not 100% in love with the flow/wording of this draft,
but I've read it a few times and I think it's ready to
share. I suppose I can always edit it later.