blog: Change listing format to -tag list (grid)
The bullet-point listing was becoming very cluttered as I added more
posts! I took a page out of danluu.com and trimmed the longer
(MM/DD/YYYY) datestamps into MM/YY (the full date is still on the page!)
and tossed in some CSS grid magic to make sure the list is actually true
to the mdoc -tag equivalent!

The trickiest part here (besides getting it to look how I wanted) was
revamping the genatom.sh to no longer parse the dates from blog.7. This
is a net-win in terms of maintainability in my book though! I'd rather
grab them from the source-of-truth .Dd macros :)