gopher/blog: Merge gopher build recipe, announce on blog
So, if you can't tell by the slew of commits leading up to this one, the
gopher commits were *actually* in their own private repo. As discussed
in 009, I decided to merge it with alexkarle.com, and this commit
represents that.

Of course, that merge also warrants a post (and the deprecation of
text.alexkarle.com). So this commit bundles in the phlog and blog
announcements (can you tell which took 1hr+ to write and which took 5m?
:P -- here's to a brave new world of micro-phlogging).