theme: Go full Acme/Plan 9
A few commits ago, I explored the idea of using Acme's cream
colored background for my site. It didn't feel quite right,
missing the other elements of the Acme UI. So I settled on
just using it for code blocks.

It was a fun challenge writing the CSS to mock the entire Acme
window (tag and all), and I'm really happy with how it turned
out. I feel like I've spent enough hours in Acme to claim to
be a fan, and my site _did_ need more color than the black on
white.